# client/swagger.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Paddokk API",
    "description": "The automotive journey sharing platform API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:7015/"
    }
  ],
  "paths": {
...
    "/api/v1/Cars/{carId}": {
      "get": {
        "tags": [
          "Cars"
        ],
        "summary": "Returns a specific car by ID — publicly accessible",
        "operationId": "Cars_GetCar",
        "parameters": [
          {
            "name": "carId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserCarDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCarDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCarDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error - unexpected failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found - the requested resource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [ ]
      }
    },
...
    "/api/v1/users/me/cars/{carId}": {
      "get": {
        "tags": [
          "UserCars"
        ],
        "summary": "Get a specific car for the current user",
        "operationId": "UserCars_GetUserCar",
        "parameters": [
          {
            "name": "carId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserCarDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCarDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCarDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error - unexpected failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found - the requested resource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - valid token but insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "UserCars"
        ],
        "summary": "Update a car for the current user",
        "operationId": "UserCars_UpdateUserCar",
        "parameters": [
          {
            "name": "carId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserCarCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserCarCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserCarCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserCarDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCarDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCarDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error - unexpected failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation or business rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found - the requested resource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - valid token but insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "UserCars"
        ],
        "summary": "Delete a car for the current user",
        "operationId": "UserCars_DeleteUserCar",
        "parameters": [
          {
            "name": "carId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal server error - unexpected failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation or business rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found - the requested resource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - valid token but insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/cars/{carId}/like": {
      "post": {
        "tags": [
          "UserCars"
        ],
        "summary": "Like a car",
        "operationId": "UserCars_LikeUserCar",
        "parameters": [
          {
            "name": "carId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error - unexpected failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation or business rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found - the requested resource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - valid token but insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "UserCars"
        ],
        "summary": "Unlike a car",
        "operationId": "UserCars_UnlikeUserCar",
        "parameters": [
          {
            "name": "carId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error - unexpected failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation or business rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found - the requested resource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - valid token but insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/cars/{carId}/subscribe": {
      "post": {
        "tags": [
          "UserCars"
        ],
        "summary": "Subscribe to a car for updates",
        "operationId": "UserCars_SubscribeToUserCar",
        "parameters": [
          {
            "name": "carId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error - unexpected failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation or business rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found - the requested resource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - valid token but insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "UserCars"
        ],
        "summary": "Unsubscribe from a car",
        "operationId": "UserCars_UnsubscribeFromUserCar",
        "parameters": [
          {
            "name": "carId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error - unexpected failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation or business rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found - the requested resource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - valid token but insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/Users/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user profile by ID",
        "operationId": "Users_GetUserById",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error - unexpected failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found - the requested resource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [ ]
      }
    },
...
    "/api/v1/Users/by-username/{username}/cars/{slug}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get a user's car by slug",
        "operationId": "Users_GetUserCarBySlug",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserCarDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCarDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCarDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error - unexpected failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found - the requested resource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [ ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "CarDriveType": {
        "type": "integer"
      },
...
      "CarSpecCategoryDto": {
        "required": [
          "category",
          "items"
        ],
        "type": "object",
        "properties": {
          "category": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "SubscriptionTier": {
        "type": "integer"
      },
...
      "UpdateUserCarCommand": {
        "required": [
          "carId",
          "customBuildName",
          "nickname",
          "color",
          "region",
          "drive",
          "engine",
          "odometerKm",
          "ownerNote",
          "specsByCategory",
          "isPrimary"
        ],
        "type": "object",
        "properties": {
          "carId": {
            "type": "integer",
            "format": "int32"
          },
          "customBuildName": {
            "type": [
              "null",
              "string"
            ]
          },
          "nickname": {
            "type": [
              "null",
              "string"
            ]
          },
          "color": {
            "type": [
              "null",
              "string"
            ]
          },
          "region": {
            "type": [
              "null",
              "string"
            ]
          },
          "drive": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CarDriveType"
              }
            ]
          },
          "engine": {
            "type": [
              "null",
              "string"
            ]
          },
          "odometerKm": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "ownerNote": {
            "type": [
              "null",
              "string"
            ]
          },
          "specsByCategory": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/CarSpecCategoryDto"
            }
          },
          "isPrimary": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
...
      "UserCarDto": {
        "required": [
          "id",
          "principalId",
          "ownerUsername",
          "slug",
          "isPublic",
          "isCustomBuild",
          "specsByCategory",
          "isPrimary",
          "createdAt",
          "updatedAt",
          "journeyCount",
          "likeCount",
          "subscriberCount",
          "isLiked",
          "isSubscribed",
          "isOwner"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "principalId": {
            "type": "string"
          },
          "ownerUsername": {
            "type": "string"
          },
          "ownerAvatarUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "slug": {
            "type": "string"
          },
          "isPublic": {
            "type": "boolean"
          },
          "carMakeId": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "carMakeName": {
            "type": [
              "null",
              "string"
            ]
          },
          "carModelId": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "carModelName": {
            "type": [
              "null",
              "string"
            ]
          },
          "carGenerationId": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "carGenerationName": {
            "type": [
              "null",
              "string"
            ]
          },
          "year": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "isCustomBuild": {
            "type": "boolean"
          },
          "customBuildName": {
            "type": [
              "null",
              "string"
            ]
          },
          "nickname": {
            "type": [
              "null",
              "string"
            ]
          },
          "color": {
            "type": [
              "null",
              "string"
            ]
          },
          "region": {
            "type": [
              "null",
              "string"
            ]
          },
          "drive": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CarDriveType"
              }
            ]
          },
          "engine": {
            "type": [
              "null",
              "string"
            ]
          },
          "odometerKm": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "ownerNote": {
            "type": [
              "null",
              "string"
            ]
          },
          "specsByCategory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CarSpecCategoryDto"
            }
          },
          "primaryImageUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "isPrimary": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "journeyCount": {
            "type": "integer",
            "format": "int32"
          },
          "likeCount": {
            "type": "integer",
            "format": "int32"
          },
          "subscriberCount": {
            "type": "integer",
            "format": "int32"
          },
          "isLiked": {
            "type": "boolean"
          },
          "isSubscribed": {
            "type": "boolean"
          },
          "isOwner": {
            "type": "boolean"
          }
        }
      },
      "UserDto": {
        "required": [
          "id",
          "email",
          "firstName",
          "username",
          "displayName",
          "createdAt",
          "emailConfirmed",
          "subscriptionTier",
          "carCount",
          "journeyCount",
          "maxCars"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": [
              "null",
              "string"
            ]
          },
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "bio": {
            "type": [
              "null",
              "string"
            ]
          },
          "avatarUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "emailConfirmed": {
            "type": "boolean"
          },
          "subscriptionTier": {
            "$ref": "#/components/schemas/SubscriptionTier"
          },
          "subscriptionExpiresAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "defaultActiveJourneyId": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "carCount": {
            "type": "integer",
            "format": "int32"
          },
          "journeyCount": {
            "type": "integer",
            "format": "int32"
          },
          "maxCars": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header. Enter your token below.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}